[fix] csv파일에 경도 위도 필드 추가
